9624: Move components into their own subdirs
