+++ app/javascript/components/UserAssignmentForm/UserAssignmentForm.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import UserAssignmentFormField from './UserAssignmentFormField';
+import UserAssignmentFormField from '../UserAssignmentFormField/UserAssignmentFormField';
 
 /**
  * User assignments form in edit user in admin mode.
